+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/WorkspaceArchiveProcessor.java
@@ -21,16 +21,10 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
  */
 
 import com.walmartlabs.concord.common.IOUtils;
-import com.walmartlabs.concord.server.jooq.enums.RawPayloadMode;
-import com.walmartlabs.concord.server.org.ResourceAccessLevel;
-import com.walmartlabs.concord.server.org.project.ProjectAccessManager;
-import com.walmartlabs.concord.server.org.project.ProjectDao;
-import com.walmartlabs.concord.server.org.project.ProjectEntry;
 import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.ProcessException;
 import com.walmartlabs.concord.server.process.ProcessKey;
 import com.walmartlabs.concord.server.process.logs.LogManager;
-import com.walmartlabs.concord.server.user.UserManager;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -39,7 +33,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
-import java.util.UUID;
 
 /**
  * Unpacks payload's workspace file, parses request data.
@@ -48,20 +41,11 @@ import java.util.UUID;
 public class WorkspaceArchiveProcessor implements PayloadProcessor {
 
     private final LogManager logManager;
-    private final ProjectDao projectDao;
-    private final ProjectAccessManager projectAccessManager;
-    private final UserManager userManager;
 
     @Inject
-    public WorkspaceArchiveProcessor(LogManager logManager,
-                                     ProjectDao projectDao,
-                                     ProjectAccessManager projectAccessManager,
-                                     UserManager userManager) {
+    public WorkspaceArchiveProcessor(LogManager logManager) {
 
         this.logManager = logManager;
-        this.projectDao = projectDao;
-        this.projectAccessManager = projectAccessManager;
-        this.userManager = userManager;
     }
 
     @Override
@@ -72,8 +56,6 @@ public class WorkspaceArchiveProcessor implements PayloadProcessor {
         if (archive == null) {
             return chain.process(payload);
         }
-
-        assertAcceptsRawPayload(payload);
 
         if (!Files.exists(archive)) {
             logManager.error(processKey, "No input archive found: " + archive);
@@ -90,47 +72,5 @@ public class WorkspaceArchiveProcessor implements PayloadProcessor {
 
         payload = payload.removeAttachment(Payload.WORKSPACE_ARCHIVE);
         return chain.process(payload);
-    }
-
-    private void assertAcceptsRawPayload(Payload payload) {
-        UUID projectId = payload.getHeader(Payload.PROJECT_ID);
-
-        if (!isRawPayloadAllowed(payload)) {
-            throw new ProcessException(payload.getProcessKey(), "The project is not accepting raw payloads: " + projectId,
-                    Status.BAD_REQUEST);
-        }
-    }
-
-    private boolean isRawPayloadAllowed(Payload payload) {
-        UUID projectId = payload.getHeader(Payload.PROJECT_ID);
-        if (projectId == null) {
-            return true;
-        }
-
-        ProjectEntry p = projectDao.get(projectId);
-        if (p == null) {
-            throw new ProcessException(payload.getProcessKey(), "Project not found: " + projectId);
-        }
-
-        RawPayloadMode m = p.getRawPayloadMode();
-        switch (m) {
-            case DISABLED: {
-                return false;
-            }
-            case OWNERS: {
-                return projectAccessManager.hasAccess(p, ResourceAccessLevel.OWNER, false);
-            }
-            case TEAM_MEMBERS: {
-                return projectAccessManager.isTeamMember(p.getId());
-            }
-            case ORG_MEMBERS: {
-                return userManager.isInOrganization(p.getOrgId());
-            }
-            case EVERYONE: {
-                return true;
-            }
-            default:
-                throw new IllegalArgumentException("Unsupported raw payload mode: " + m);
-        }
     }
 }